requirements.txt
asgiref==3.8.1
boto3==1.36.22
botocore==1.36.22
cffi==1.17.1
cryptography==44.0.1
Django==5.1.6
django-cors-headers==4.7.0
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
jmespath==1.0.1
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0

pandas==2.2.3
xlsxwriter==3.2.2

requests==2.31.0
